+++ test/exhaustive/agescx.test.ts
@@ -1,12 +1,12 @@
 import { readFileSync } from "fs";
-import { scrap, list, text, link } from "../../lib";
+import { scrap, list, text, link, html } from "../../lib";
 
-const html = readFileSync('./test/exhaustive/agescx.html').toString();
+const shtml = readFileSync('./test/exhaustive/agescx.html').toString();
 
 describe('agescx documentation', () => {
 
     it('should get all navigation items from agescx', () => {
-        const result = scrap(html, {
+        const result = scrap(shtml, {
             navs: list('ul.navbar-nav>li:not(.disabled)', {
                 text: text('li > a'),
                 link: link('li > a'),
@@ -14,9 +14,16 @@ describe('agescx documentation', () => {
                     text: text('li > a'),
                     link: link('li > a')
                 })
-            })
+            }),
+            title: text('h1'),
+            content: html('div[role="main"]')
         });
+        
         expect(result.navs.length).toBe(9);
+        expect(result.navs[2].submenu.length).toBe(2);
+        expect(result.navs[2].submenu[0].text).toBe('Adding new unit');
+        expect(result.content.length).toBeGreaterThan(30);
+        expect(result.title).toBe('Agescx Documentation');
     });
 
 });